Refactor and expand Obsidian API, remove legacy code
Major refactor introducing new modules for Crucible and Nexo APIs, adding scripting and GUI support, and expanding item/block functionality. Removes legacy and unused classes, including obsolete enchantment and shape-related code. Updates existing modules for improved modularity and maintainability.

+++ src/main/java/org/quiltmc/qsl/fluid/mixin/EntityMixin.java
@@ -41,7 +41,7 @@ import javax.annotation.Nullable;
 @Mixin(Entity.class)
 public abstract class EntityMixin implements CustomFluidInteracting {
 	@Shadow
-	public float fallDistance;
+	public double fallDistance;
 	@Shadow
 	private Level level;
 	@Shadow
